requirements.txt
loguru
black
openai
pydantic
pytest
python-dotenv
requests
#langchain
#streamlit
pyTelegramBotAPI
#fastapi
